+++ src/babashka/impl/clojure/java/browse.clj
@@ -19,13 +19,14 @@
           :windows)))
 
 (defn browse-url [url]
-  (let [url (str url)]
-    (if-let [script (-> open-url-script deref deref)]
-      (sh script url)
-      (case os
-        :mac (sh "open" url)
-        :linux (sh "xdg-open" url)
-        :windows (sh "cmd" "/C" "start" (.replace url "&" "^&"))))))
+  (let [url (str url)
+        {:keys [exit]} (if-let [script (-> open-url-script deref deref)]
+                         (sh script url)
+                         (case os
+                           :mac (sh "open" url)
+                           :linux (sh "xdg-open" url)
+                           :windows (sh "cmd" "/C" "start" (.replace url "&" "^&"))))]
+    (zero? exit)))
 
 (def browse-namespace
   {'*open-url-script* open-url-script