+++ src/babashka/impl/clojure/main.clj
@@ -53,6 +53,7 @@
 (def ^{:doc "A sequence of lib specs that are applied to `require`
 by default when a new command-line REPL is started."} repl-requires
   '[[clojure.repl :refer (source apropos pst dir doc find-doc)]
+    [clojure.java.javadoc :refer (javadoc)]
     [clojure.pprint :refer (pp pprint)]])
 
 (defmacro with-read-known